ingestao-mongo/requirements.txt
pymongo==4.9.0
pandas==2.2.2
pyarrow==17.0.0
google-cloud-storage==2.18.2
google-cloud-secret-manager==2.20.2
google-cloud-bigquery==3.25.0
python-dotenv==1.0.1